Create app factory
