+++ orca/wsgi.py
@@ -1,0 +1,8 @@
+from orca import create_app
+
+
+app = create_app()
+
+
+if __name__ == "__main__":
+    app.run(host="0.0.0.0")
